% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.r
\name{dem.add}
\alias{dem.add}
\title{Construct new Variables}
\usage{
dem.add(x, what, id. = id, breaks = c(3, 15, 20, 30, 40, 50, 60))
}
\arguments{
\item{x}{A data.tabel, typically \code{dem}.}

\item{what}{A length one character vector. What set of variabels? Corrently 
available are: \code{'rim','calendar','ageclass','gm','zone','lifestage'},
see details.}

\item{id.}{The \code{id} object. See \code{id <- setup()}.}

\item{breaks}{An optional numeric vector, sepcifying the age breaks to split 
age into ageclasses. Default are the 7-folded ageclasses used to weight the 
TV-panel: \code{c(3,15,20,30,40,50,60)}.}
}
\value{
\code{x}, the input data.table (usually \code{dem}) is changed in 
place. The new variabels are added as new columns to \code{x} and the 
corresponding variable values and labels are added to \code{id$lab}.
}
\description{
\code{dem.add} is a generic function. It is intended to by applied to the 
\code{dem} object and constructs new variables based on existing ones. The 
new variables are added as columns to dem and the labels to the variable 
values are added to \code{id$lab}.
}
\details{
To allow the construction of new variabels based on other variabels, these 
other variabels have to be present in the \code{x} (usually \code{dem}). This 
means they have to be imported in the first place, e.g.: 
\code{setup(..., dem.var = c('v1','v2'))}. How do we know which variables are 
needed? There is a parameter \code{need} in each subfunction where the needed 
variabels are listed. So, to add rims first ask what variabels are needed 
and the import the data accordingly, see examples.

Careful with \code{'gm'} and \code{'zone'}. These read the corresponding CSV 
file in ../data/meta/.
}
\examples{
library(tv)
dem.var <- eval(formals(tv:::dem.add.rim)$need)
setup('2018-01-01', 3, dem.var = dem.var, import = TRUE, view = FALSE, prg = FALSE)
dem[]               # dem contains additional variables specified in dem.var
dem.add(dem, 'rim') # construct new variables 'rim' (weighting rims).
dem[]               # The new variables are added to dem
id$lab$rim_agesex   # And theis labels are added to id$lab

# dem.add.calendar only needs 'day' which is always present:
setup('2018-01-01', 3, import = TRUE, view = FALSE, prg = FALSE, dem.uni = FALSE)
dem.add(dem, 'calendar')
id$lab$month

}
