% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.r
\name{dem.rim}
\alias{dem.rim}
\title{import dem and add rim variables
this is simply a wrapper for setup() but handles problems in old .dem files}
\usage{
dem.rim(from, to)
}
\description{
import dem and add rim variables
this is simply a wrapper for setup() but handles problems in old .dem files
}
