% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.r
\name{ipf}
\alias{ipf}
\title{iterative proportional fitting}
\usage{
ipf(data, rims, crit = 1e-04, iter = 250)
}
\value{
a list containing a numeric vector of the weights and a vector of 
the convergence criterion for each iteration
}
\description{
iterative proportional fitting
}
