% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.r
\name{overlap}
\alias{overlap}
\title{these functions together solve the task of merging a lookup table to viewing 
by start end interval. This is relevant for timeband and programs. The 
workorse is data.table::foverlaps() in overlap()}
\usage{
overlap(x, y, keys = NULL, nomatch = NA)
}
\description{
these functions together solve the task of merging a lookup table to viewing 
by start end interval. This is relevant for timeband and programs. The 
workorse is data.table::foverlaps() in overlap()
}
