% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prg.r
\name{read.prg}
\alias{read.prg}
\title{Reads the .std-file}
\usage{
read.prg(id)
}
\description{
The .std file contains the program logs ("Sendeprotokoll") for a given day.
}
\details{
The program logs are stored in ".std" files. For each day there is a separate 
 file. The files are of fixed-with type, with up to 55 fields, some fields as
 long as 100 characters. Following are the first fields. Possably not all of 
 these fields are needed, e.g 3, 6, 7, 8 and 9 might be not needed:

field	 name            label	 length class
 1	     channel code    sta	   5	    integer (the code of virtual channels)
 2	     BrdCstId	       prg	   8	    integer (broadcast id)
 3	     date	           date	   8	     integer (is equal to file date)
 4	     start	         start	 6	    integer
 5	     duration	       dur	   5	    integer
 6	     BrdCstTimePlan  ref	  15      integer (date/time reference for some ads)
 7	     Title	         -	   100	    character
 8	     Description	   -	   100	    character
 9       Comments        -     100      character
10       BrdCstTyp       typ     2      integer (Senderuebergreifende definition)
..       ...             ...     .      ...
24       BrdCstSeq       seq     3      integer
25       SumPieces       seqN    3      integer
..       ...             ...     .      ...

Inkonsistenzen
- DE180224.STD genre code (2 Zeichen lang) auf Zeile 1608 ist "1.", read.prg
  gibt eine warnung aus, leist das File aber korrekt, mit coercion to NA
}
