% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.r
\name{sum.weight}
\alias{sum.weight}
\title{caclulate universe approximately by summing up the weights for each rim and 
each day. Guest are excluded, they are not part of the universe. The weight 
of a guests is always a copy of the weight of the housewive.
Value as well as labels are returned because values need to match the values 
in margins for ipf weighting. Labels allow to check that the same value have
also the same meaning}
\usage{
\method{sum}{weight}(dem, cols = grep("rim", names(dem), value = TRUE),
  w = "weight")
}
\description{
caclulate universe approximately by summing up the weights for each rim and 
each day. Guest are excluded, they are not part of the universe. The weight 
of a guests is always a copy of the weight of the housewive.
Value as well as labels are returned because values need to match the values 
in margins for ipf weighting. Labels allow to check that the same value have
also the same meaning
}
