% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dt.r
\name{na.to.0}
\alias{na.to.0}
\alias{ordercol}
\alias{id.dt}
\alias{id.dt.cross}
\title{utility functions for R data.table}
\usage{
na.to.0(dt, to = 0L)

ordercol(dt, cols, nextto = 1L)

id.dt(label, id.name = "id", id = NULL, label.name = "label",
  roll = NULL)

id.dt.cross(x)
}
\description{
utility functions for R data.table
}
