% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tv.r
\name{id.time}
\alias{id.time}
\alias{id.tmb}
\alias{id.day}
\alias{pin.type}
\title{Utility functions for the tv package}
\usage{
id.time()

id.tmb(x)

id.day(day, to, dem.day = NULL)

pin.type(dt)
}
\description{
These utility functions are very specific for the tv rawdata.
}
\details{
\code{id.tmb}  reshapes the timeband object in 
\code{setup(tmb = list('wholeday' = c(start = '02:00:00', end = '25:59:59')))}
(default) into a more compact form for use in \code{overlap.join()}. Gives 
labels for each \code{tmb} if not specified in \code{names(tmb)}.
Important: if 'end' is specified like in Instar, than 1 second has to be 
subtracted to match Instar results! 
E.g.: if specified in Instar: 08:03:07 - 12:04:25 the equivalent for this
      tv package is \code{c(start = 08:03:07, end = 12:04:24)}, that means: end-1

\code{id.day} is a function to conveniently create a sequence of Dates.
For convenience there are three ways of specifying the days to analyse:
\enumerate{
  \item by entering a prespecified vector of dates in \code{'day'}.
  \item by specifying the start end end dates in \code{'day'} and \code{'to'} respectively.
  \item by specifying start date \code{'day'} and the length of the sequence of days
  in \code{'nday'}. If \code{'nday'} is negative the sequence will be created backwards.
}
day is a character vector of dates in the form '2015-12-31'
Time shifted viewing is limited to 7 days by definition.
The demografics file is usually loaded for the live dates but this is 
allowed to differ: Analysing the viewing of a fixed sample at one
day alows to circumvent the fact that otherwise weights change every day.
use class Date to correctly creating sequencies of dates but store dates as 
character, its much easier to program with and more efficient
}
